requirements.txt
alembic==0.9.3
aniso8601==1.2.1
APScheduler==3.3.1
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-APScheduler==1.7.0
Flask-HTTPAuth==3.2.3
Flask-Migrate==2.0.4
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
jpush==3.2.8
Mako==1.0.7
MarkupSafe==1.0
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
requests==2.18.1
six==1.10.0
SQLAlchemy==1.1.11
tzlocal==1.4
urllib3==1.21.1
uWSGI==2.0.15
Werkzeug==0.12.2
lxml==3.8.0
pillow
pytesseract